order of widgets to float to allow tweaking

+++ beegisplugins/eu.hydrologis.jgrass.formeditor/src/eu/hydrologis/jgrass/featureeditor/xml/annotated/ALabel.java
@@ -51,7 +51,7 @@ public class ALabel extends FormElement {
      * The widget order.
      */
     @XmlAttribute(name = ORDER)
-    public Integer order = null;
+    public Float order = null;
 
     /**
      * The layout constraints.
@@ -60,9 +60,9 @@ public class ALabel extends FormElement {
     public String constraints = null;
 
     @Override
-    public int getOrder() {
+    public float getOrder() {
         if (order == null) {
-            order = 0;
+            order = 0f;
         }
         return order;
     }
