+++ beegisplugins/eu.hydrologis.jgrass.formeditor/src/eu/hydrologis/jgrass/featureeditor/xml/annotated/ARadioButton.java
@@ -81,7 +81,7 @@ public class ARadioButton extends FormElement {
      * The widget order.
      */
     @XmlAttribute(name = ORDER)
-    public Integer order = null;
+    public Float order = null;
 
     /**
      * The layout constraints.
@@ -90,9 +90,9 @@ public class ARadioButton extends FormElement {
     public String constraints = null;
 
     @Override
-    public int getOrder() {
+    public float getOrder() {
         if (order == null) {
-            order = 0;
+            order = 0f;
         }
         return order;
     }
